refactor TraceActorContext
