+++ src/tools/replay/nodes/TraceActorCreationNode.java
@@ -3,6 +3,7 @@ package tools.replay.nodes;
 import tools.concurrency.TracingActors.TracingActor;
 import tools.replay.actors.ActorExecutionTrace;
 import tools.replay.actors.ActorExecutionTrace.ActorTraceBuffer;
+import tools.replay.nodes.TraceContextNodes.TraceActorContextNode;
 
 
 public final class TraceActorCreationNode extends TraceNode {
