+++ src/tools/replay/nodes/TraceMessageNode.java
@@ -14,6 +14,7 @@ import tools.replay.actors.ActorExecutionTrace.ActorTraceBuffer;
 import tools.replay.actors.ExternalEventualMessage.ExternalDirectMessage;
 import tools.replay.actors.ExternalEventualMessage.ExternalPromiseCallbackMessage;
 import tools.replay.actors.ExternalEventualMessage.ExternalPromiseSendMessage;
+import tools.replay.nodes.TraceContextNodes.TraceActorContextNode;
 
 
 public abstract class TraceMessageNode extends TraceNode {
